+++ src/main/java/backend/like_house/domain/user/converter/UserConverter.java
@@ -1,25 +1,24 @@
 package backend.like_house.domain.user.converter;
 
-import backend.like_house.domain.user.dto.SignInDTO;
-import backend.like_house.domain.user.dto.SignUpDTO;
+import backend.like_house.domain.user.dto.UserDTO;
 import backend.like_house.domain.user.entity.User;
 
 public class UserConverter {
 
-    public static SignUpDTO.SignUpResponse toSignUpResponseDTO(User user) {
-        return SignUpDTO.SignUpResponse.builder()
+    public static UserDTO.SignUpResponse toSignUpResponseDTO(User user) {
+        return UserDTO.SignUpResponse.builder()
                 .name(user.getName())
                 .email(user.getEmail())
                 .build();
     }
 
-    public static SignInDTO.SignInResponse toSignInResponseDTO(String accessToken) {
-        return SignInDTO.SignInResponse.builder()
+    public static UserDTO.SignInResponse toSignInResponseDTO(String accessToken) {
+        return UserDTO.SignInResponse.builder()
                 .accessToken(accessToken)
                 .build();
     }
 
-    public static User toUser(SignUpDTO.SignUpRequest signUpRequest, String password) {
+    public static User toUser(UserDTO.SignUpRequest signUpRequest, String password) {
         return User.builder()
                 .name(signUpRequest.getName())
                 .email(signUpRequest.getEmail())
